requirements.txt
altgraph==0.15
awscli==1.14.50
boto3==1.6.2
botocore==1.9.3
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.7
cycler==0.10.0
dj-database-url==0.4.2
django==2.0
django-cors-headers==2.1.0
django-heroku==0.2.0
django-oauth-toolkit==1.0.0
django-storages==1.6.5
django-webpack-loader==0.5.0
djangorestframework==3.7.7
docutils==0.14
future==0.16.0
gunicorn==19.7.1
httpie==0.9.9
idna==2.6
jmespath==0.9.3
kiwisolver==1.0.1
macholib==1.9
matplotlib==2.2.2
nltk==3.2.5
numpy==1.14.2
oauthlib==2.0.7
pefile==2017.11.5
pew==1.1.2
Pillow==5.0.0
pipenv==10.1.0
psycopg2==2.7.3.2
pyasn1==0.4.2
Pygments==2.2.0
PyInstaller==3.3.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.2.6
whitenoise==3.3.1
